Change main package name
