+++ src/main/java/com/bring/library/App.java
@@ -1,0 +1,6 @@
+package com.bring.library;
+
+public class App {
+    public static void main(String[] args) {
+    }
+}
