+++ src/main/java/com/bring/library/annotation/Configuration.java
@@ -1,0 +1,4 @@
+package com.bring.library.annotation;
+
+public @interface Configuration {
+}
