+++ src/main/java/com/bring/library/annotation/Dao.java
@@ -1,0 +1,11 @@
+package com.bring.library.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Dao {
+}
