+++ src/main/java/com/bring/library/configuration/AppConfiguration.java
@@ -1,0 +1,4 @@
+package com.bring.library.configuration;
+
+public class AppConfiguration {
+}
