+++ src/main/java/com/bring/library/exception/BoboNotFoundException.java
@@ -1,0 +1,4 @@
+package com.bring.library.exception;
+
+public class BoboNotFoundException {
+}
