+++ src/main/java/com/bring/library/exception/ClassHasMoreOneInterfacesException.java
@@ -1,0 +1,7 @@
+package com.bring.library.exception;
+
+public class ClassHasMoreOneInterfacesException extends RuntimeException{
+    public ClassHasMoreOneInterfacesException(String message) {
+        super(message);
+    }
+}
