+++ src/main/java/com/bring/library/exception/CreateInstanceException.java
@@ -1,0 +1,7 @@
+package com.bring.library.exception;
+
+public class CreateInstanceException extends RuntimeException {
+    public CreateInstanceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
