+++ src/main/java/com/bring/library/exception/InvalidFileNameException.java
@@ -1,0 +1,7 @@
+package com.bring.library.exception;
+
+public class InvalidFileNameException extends RuntimeException {
+    public InvalidFileNameException(String message) {
+        super(message);
+    }
+}
