+++ src/main/java/com/bring/library/exception/SetValueToFieldException.java
@@ -1,0 +1,7 @@
+package com.bring.library.exception;
+
+public class SetValueToFieldException extends RuntimeException {
+    public SetValueToFieldException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
