+++ src/main/java/com/bring/library/registry/ApplicationRegistry.java
@@ -1,0 +1,4 @@
+package com.bring.library.registry;
+
+public class ApplicationRegistry {
+}
