+++ src/main/java/com/bring/library/service/creator/BoboCreatorService.java
@@ -1,0 +1,4 @@
+package com.bring.library.service.creator;
+
+public interface BoboCreatorService {
+}
