+++ src/main/java/com/bring/library/service/processor/AnnotationProcessorService.java
@@ -1,0 +1,4 @@
+package com.bring.library.service.processor;
+
+public interface AnnotationProcessorService {
+}
