+++ src/main/java/com/bring/library/service/processor/impl/AnnotationProcessorServiceImpl.java
@@ -1,0 +1,6 @@
+package com.bring.library.service.processor.impl;
+
+import com.bring.library.service.processor.AnnotationProcessorService;
+
+public class AnnotationProcessorServiceImpl implements AnnotationProcessorService {
+}
