+++ src/main/java/com/bring/library/util/AnnotationProcessorUtil.java
@@ -1,0 +1,4 @@
+package com.bring.library.util;
+
+public class AnnotationProcessorUtil {
+}
