--- src/main/java/com/bring/project/annotation/Configuration.java
@@ -1,4 +1,0 @@
-package com.bring.project.annotation;
-
-public @interface Configuration {
-}
