--- src/main/java/com/bring/project/configuration/AppConfiguration.java
@@ -1,4 +1,0 @@
-package com.bring.project.configuration;
-
-public class AppConfiguration {
-}
