--- src/main/java/com/bring/project/exception/BoboNotFoundException.java
@@ -1,4 +1,0 @@
-package com.bring.project.exception;
-
-public class BoboNotFoundException {
-}
