--- src/main/java/com/bring/project/exception/ClassHasMoreOneInterfacesException.java
@@ -1,7 +1,0 @@
-package com.bring.project.exception;
-
-public class ClassHasMoreOneInterfacesException extends RuntimeException{
-    public ClassHasMoreOneInterfacesException(String message) {
-        super(message);
-    }
-}
