--- src/main/java/com/bring/project/exception/CreateInstanceException.java
@@ -1,7 +1,0 @@
-package com.bring.project.exception;
-
-public class CreateInstanceException extends RuntimeException {
-    public CreateInstanceException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
