--- src/main/java/com/bring/project/exception/EmptyClassLoaderException.java
@@ -1,7 +1,0 @@
-package com.bring.project.exception;
-
-public class EmptyClassLoaderException extends RuntimeException {
-    public EmptyClassLoaderException(String message) {
-        super(message);
-    }
-}
