--- src/main/java/com/bring/project/exception/EmptyDirectoryException.java
@@ -1,7 +1,0 @@
-package com.bring.project.exception;
-
-public class EmptyDirectoryException extends RuntimeException{
-    public EmptyDirectoryException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
