--- src/main/java/com/bring/project/exception/InvalidFileNameException.java
@@ -1,7 +1,0 @@
-package com.bring.project.exception;
-
-public class InvalidFileNameException extends RuntimeException {
-    public InvalidFileNameException(String message) {
-        super(message);
-    }
-}
