--- src/main/java/com/bring/project/exception/NotFoundAnnotationException.java
@@ -1,7 +1,0 @@
-package com.bring.project.exception;
-
-public class NotFoundAnnotationException extends RuntimeException {
-    public NotFoundAnnotationException(String message) {
-        super(message);
-    }
-}
