--- src/main/java/com/bring/project/exception/SetValueToFieldException.java
@@ -1,7 +1,0 @@
-package com.bring.project.exception;
-
-public class SetValueToFieldException extends RuntimeException {
-    public SetValueToFieldException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
