--- src/main/java/com/bring/project/registry/ApplicationRegistry.java
@@ -1,4 +1,0 @@
-package com.bring.project.registry;
-
-public class ApplicationRegistry {
-}
