--- src/main/java/com/bring/project/service/creator/BoboCreatorService.java
@@ -1,4 +1,0 @@
-package com.bring.project.service.creator;
-
-public interface BoboCreatorService {
-}
