--- src/main/java/com/bring/project/service/processor/AnnotationProcessorService.java
@@ -1,4 +1,0 @@
-package com.bring.project.service.processor;
-
-public interface AnnotationProcessorService {
-}
