--- src/main/java/com/bring/project/util/AnnotationProcessorUtil.java
@@ -1,4 +1,0 @@
-package com.bring.project.util;
-
-public class AnnotationProcessorUtil {
-}
